requirements-prod.txt
# Production dependencies
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
redis==6.4.0
alembic==1.14.0
pydantic==2.11.7
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.18
pika==1.3.2
websockets==12.0
pyyaml==6.0.2
